+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Services/TrustOverviewServiceModelTests.cs
@@ -7,7 +7,7 @@ public class TrustOverviewServiceModelTests
 {
     private static readonly TrustOverviewServiceModel BaseTrustOverviewServiceModel =
         new("1234", "", "", "", TrustType.MultiAcademyTrust, "", "", null, null, 0, new Dictionary<string, int>(), 0,
-            0);
+            0, false);
 
     [Theory]
     [InlineData(100, 100, 100)]
@@ -18,13 +18,14 @@ public class TrustOverviewServiceModelTests
     [InlineData(100, 30, 30)]
     [InlineData(4, 1, 25)]
     [InlineData(3, 1, 33)]
-    public void PercentageFull_ReturnsCorrectValue(int totalCapacity, int totalPupilNumbers, int? expectedPercentage)
+    public void PercentageFull_ReturnsCorrectValue_whenCapacityDataComplete(int totalCapacity, int totalPupilNumbers, int? expectedPercentage)
     {
         // Arrange
         var model = BaseTrustOverviewServiceModel with
         {
             TotalCapacity = totalCapacity,
-            TotalPupilNumbers = totalPupilNumbers
+            TotalPupilNumbers = totalPupilNumbers,
+            HasIncompleteCapacityData = false
         };
 
         // Act
@@ -32,5 +33,23 @@ public class TrustOverviewServiceModelTests
 
         // Assert
         percentageFull.Should().Be(expectedPercentage);
+    }
+
+    [Fact]
+    public void PercentageFull_ReturnsNull_whenCapacityDataIncomplete()
+    {
+        // Arrange
+        var model = BaseTrustOverviewServiceModel with
+        {
+            TotalCapacity = 100,
+            TotalPupilNumbers = 100,
+            HasIncompleteCapacityData = true
+        };
+
+        // Act
+        var percentageFull = model.PercentageFull;
+
+        // Assert
+        percentageFull.Should().BeNull();
     }
 }